requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
appdirs==1.4.4            # via -r requirements/base.txt, fs
attrs==20.2.0             # via pytest
click==7.1.2              # via code-annotations
code-annotations==0.7.0   # via -r requirements/test.in
coverage==5.3             # via pytest-cov
ddt==1.4.1                # via -r requirements/test.in
djangorestframework==3.11.1  # via -r requirements/base.txt
fs==2.4.11                # via -r requirements/base.txt, xblock
httpretty==1.0.2          # via -r requirements/test.in
importlib-metadata==1.7.0  # via pluggy, pytest
iniconfig==1.0.1          # via pytest
jinja2==2.11.2            # via code-annotations
lxml==4.5.2               # via -r requirements/base.txt, xblock
mako==1.1.3               # via -r requirements/base.txt, xblock-utils
markupsafe==1.1.1         # via -r requirements/base.txt, jinja2, mako, xblock
mock==3.0.5               # via -r requirements/test.in
more-itertools==8.5.0     # via pytest
packaging==20.4           # via pytest
pathlib2==2.3.5           # via pytest
pbr==5.5.0                # via stevedore
pillow==7.2.0             # via -r requirements/test.in
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pycodestyle==2.6.0        # via -r requirements/test.in
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.1        # via -r requirements/test.in
pytest-django==3.10.0     # via -r requirements/test.in
pytest==6.0.2             # via pytest-cov, pytest-django
python-dateutil==2.8.1    # via -r requirements/base.txt, xblock
python-slugify==4.0.1     # via code-annotations
pytz==2020.1              # via -r requirements/base.txt, django, fs, xblock
pyyaml==5.3.1             # via -r requirements/base.txt, code-annotations, xblock
simplejson==3.17.2        # via -r requirements/base.txt, xblock-utils
six==1.15.0               # via -r requirements/base.txt, fs, mock, packaging, pathlib2, python-dateutil, stevedore, xblock
sqlparse==0.3.1           # via -r requirements/base.txt, django
stevedore==1.32.0         # via code-annotations
text-unidecode==1.3       # via python-slugify
toml==0.10.1              # via pytest
typing==3.7.4.3           # via -r requirements/base.txt, fs
web-fragments==0.3.2      # via -r requirements/base.txt, xblock, xblock-utils
webob==1.8.6              # via -r requirements/base.txt, xblock
xblock-utils==2.1.1       # via -r requirements/base.txt
xblock==1.4.0             # via -r requirements/base.txt, xblock-utils
zipp==1.2.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
